requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
fastapi==0.65.2
h11==0.12.0
httptools==0.2.0
idna==2.10
importlib-metadata==4.5.0
pydantic==1.8.2
python-dotenv==0.17.1
PyYAML==5.4.1
requests==2.25.1
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
watchgod==0.7
websockets==9.1
zipp==3.4.1
zipstream-new==1.1.8
